Refactor PipePool to use isize (fd) instead of StdHandle
- PipePool now uses (Handle, isize) as key instead of (Handle, StdHandle)
- Decouples PipePool from actor_runtime::StdHandle
- pipe_path() now takes isize, generates paths like "pipes/actor-{id}-{fd}"
- Removes std_handle from MaterializedWriter (no longer needed)
- Updates all callers to pass fd as isize

+++ ailets-rs/ailetos/src/attachments.rs
@@ -113,7 +113,7 @@ impl AttachmentManager {
         let task = tokio::spawn(async move {
             // Get reader for the realized writer
             let reader = match pipe_pool
-                .get_or_await_reader((node_handle, std_handle), false, &id_gen)
+                .get_or_await_reader((node_handle, std_handle as isize), false, &id_gen)
                 .await
             {
                 Ok(reader) => reader,
